Fix Wild Arms 2 overworld forest sprite; some rearranging of code;
